Convert later pipeline stages to use new .canvasBuffer
As part of the current viewport overhaul, each pdImage is going to start
caching a full copy of its composited image.  The viewport will simply
extract the required region as necessary, but it will leave any
compositing to the parent pdImage object.

To that end, the current "back/front buffer" system is too simplified.
We now need three buffers to render the viewport efficiently:

1) compositeBuffer: a fully composited copy of the current image (and is
the same size as the image).  This has many uses outside the viewport,
obviously.
2) canvasBuffer: the relevant region of compositeBuffer, cropped and
zoomed according to the current viewport.  (This buffer is the same size
as the viewport.)
3) frontBuffer: this is stored *only by the viewport*.  It begins as a
copy of the current canvasBuffer, but then any UI elements (transform
nodes, brush outlines, etc) are composited onto it as necessary.  For
canvas MouseMove events, this buffer is the one we modify for things
like brush outlines, which makes interaction much smoother.