pdCompositor: wrap up first round of full-image compositing optimizations
The new method is now slightly faster than the old method... with an
ugly caveat: custom blend-modes or non-destructive effects murder
performance (because they have to be applied to the full-size layer, and
the code for these is pure VB).

I am starting work on a new compositor solution that will mitigate some
of these performance issues, but I don't know that it's possible to
mitigate *all* of them.  What I may end up doing is providing both
viewport pipelines in parallel, and letting the associated "viewport
rendering preference" swap between the two at the user's discretion.
TBD pending additional