pdCompositor: unify compositing results of all rendering pipeline modes
This update also fixes some issues with non-standard blend modes on
images with complicated alpha channels and highly variable colors.

Since it was convenient, I've also prepped the code to make the future
addition of alpha inheritance trivial.  Alpha inheritance is a brilliant
feature in Krita
(https://userbase.kde.org/Krita/Tutorial_2#Inherit_Alpha_.28alpha_.3D_transparency.29)
and it makes coloring work much, much simpler.  PD's compositor will be
able to do this in a few lines of code; all that's holding me back is
implementing a UI to expose the option to the user.