Wrap up this round of compositor optimizations
I'm sure there will be more to optimize in the future, but as of now
(build 70), PD's viewport pipeline is in a great place.  Everything is
more fluid and better organized, and the viewport pipeline and
compositor will only need a few lines of code to handle paint tools.

As an added bonus, paint operations will automatically receive all layer
features for free (opacity, blend modes, subpixel positioning, etc)
without any work on the part of the brush engine.  Very exciting.