Text tool: start work on bare-bones UI bits
The text tool is going to be a massive project, since text is PD's first
true vector tool.  As such, the pdLayer object is going to require many
updates.  (Obviously, once this project is done, adding future vector
tools - shapes, pen, etc - will be much easier thanks to having a
functional vector management pipeline.)

Text tool is also a UI nightmare since a lot of user-facing options are
relevant, and PD will need a dedicated edit box for text entry. (I do
not plan to implement on-canvas editing - sorry!  Blame WAPI for this.)

Special user controls like a font-specific combo box are also important,
not just here but elsewhere in the project, so that's another big
component of this work.

At first, I'm hoping to simply implement a "bare bones" version of a
text tool, to make sure my vector engine design is sound.  (I also need
to write a new GDI+ font manager, which is going to take some time.)
Once a bare bones implementation is functional and the new vector
pipeline is in place, I can dedicate a lot more time to cleaning up the
text tool UI and expanding available features.