pdTextRenderer: wrap up bulk of property get/set functionality
A class instance can now be properly serialized to/from XML, which is
the main functionality required for PDI storage and Undo/Redo
functionality.  GDI+ objects are also cached and cleaned when relevant,
and font caching has been implemented to minimize churn.

Next, it's time to work on the actual rendering bits.