New: build system-specific font cache at startup
Performance should be excellent.  On my old Win 7 PC, a system with ~250
fonts averages 12-13 ms to retrieve and sort the list of installed
fonts.  (Sorting is required because Windows returns fonts in an
arbitrary order, and we need to mirror the output to various drop-downs
throughout the program.)

(Also, this makes me wonder why GIMP's startup font detection is so
horrifically slow.)