pdComboBox_Font: start working on a font-specific drop down
I thought about working this functionality directly into pdComboBox, but
it's not worth the trouble, IMO.  A font selection dropdown is a really
odd duck for a couple reasons:

1) Its list contents are managed 100% internally, so it shouldn't expose
any add/remove/clear type functionality.
2) It has some incredibly specific rendering requirements, as we have to
create hundreds (possibly thousands) of fonts on-the-fly.  That kind of
messy work would just clutter up a normal combo box.
3) It requires a lot of interop with the separate Font_Management
module, as this class doesn't do things like query the system for fonts;
instead, it relies on the module to do that.
4) For saving/loading presets in a "portable app friendly" manner, this
class needs to store the text contents of the combo box rather than the
.ListIndex (as the availability of fonts will vary from system to
system).

Anyway, those are a few of the reasons that this is its own control.

I'll be cracking away at this over the next few weeks amidst other UI
work.