Start prep work for OpenType font conversion
This is going to be a hideously ugly project, but it's theoretically
doable.  GDI supports OpenType fonts (and has for years), so I should be
able to parse GDI glyphs, convert them from quadratic splines to GDI+
compatible Bezier curves, then import them to GDI+ as a GraphicsPath.
This would allow me to use GDI+ to render OpenType fonts, enabling all
the nice GDI+ features we have for TrueType fonts.  Fingers crossed that
it works.

This commit is mostly a bunch of refactoring to make the new behavior
easier to implement.