Finish overhaul of tool option panels
Glad to have this ugly project over with.  Instead of stuffing all tool
option panels into picture boxes, and carefully layering the picture
boxes on the option panel form, each tool panel is now its own window.
These windows are simply swapped in/out when new buttons are clicked,
and the window manager takes care of synchronizing them to the parent
form's position.

This will make it much easier to work on tool option panels in the
future.