Add new Layer > Rasterize menu
This layer entry will never be explicitly required (e.g. PD will be
smart enough to raise dialogs as necessary if action(s) require
rasterization), but this menu is helpful for debugging, and to help
users coming from Photoshop.

Actual rasterizing is not implemented yet, but it will be in the next
few days.