Implement new UNDO_LAYER_VECTORSAFE action type
Pretty much all adjustments and effects are not vector-safe, meaning
that if you apply one to a vector layer, the layer *must* be rasterized
first.

This is nice, because PD can use the default UNDO_LAYER notification to
know if an operation is not vector-safe, allowing it to raise a
rasterization warning.

If I someday offer tool variants that *are* vector-safe (e.g. an Invert
function that simply inverts vector colors instead of per-pixel
modifications), the new UNDO_LAYER_VECTORSAFE command will let the
processor know that this action does *not* require rasterization first,
so no warning is necessary.