Move accelerator enable/disable behavior directly into pdTextBox
Yep, this is ugly.  Nope, there's not a better way to do it.  Focus with
our various API-created windows is a mess, but it's crucially important
that we suspend accelerators when a text box has focus (since individual
letters like L and R trigger major events otherwise).

Hypothetically we could allow *some* accelerators to still propagate,
but I've yet to identify a right way to do this.  Until that point, I
prefer to err on the side of caution and simply ignore accelerators
until text boxes lose focus.