Implement new non-destructive event Undo/Redo handling
Relates to #169.

Most image editors only save Undo/Redo data when an entire layer loses
focus.  I'm not a fan of this, because *so many things* can change
during a single interaction with a layer.  For example, on a text layer,
you might modify a dozen settings a dozen times before switching to a
new text layer, and only being able to Undo/Redo those changes in a
single batch is just not fine-grained enough IMO.

To that end, I've now implemented a trial system where non-destructive
changes are tracked on a per-control basis.  My trial control is the
primary text box on the text tool.

When the text box gains focus, a new flagInitialNDFXState() function is
called in the central processor, with the affected setting ID and
current setting value noted.

When the control loses focus, a corresponding flagFinalNDFXState()
function is called.  This function compares the value at focus entry
with the value at focus exit, and if changes are found, Undo data is
generated.  I've written some new custom Undo/Redo interactions to make
this as quick as possible, and PD's vector layers save very rapidly as
we're only saving a compressed XML string out to file.

This technique allows me to solve a few different problems at once:

1) Fine-grained modifications are saved, without bombarding the
Undo/Redo chain with events generated by a single control (e.g. every
letter typed into a text box).  In my mind, this is the best compromise
for saving "just enough" Undo/Redo data.  (Once the Undo History is
converted to a proper collapsible section of the right-side toolbox,
this technique will be even better.)

2) The new system is integrated directly into the Macro recorder, which
should allow for macro operations like adding a Copyright text layer to
a collection of images.  (I haven't actually tested this, but "in
theory" it should work.)

3) The new code is a nice compromise between limiting the amount of code
required to check non-destructive changes, and reliably tracking changes
to the image.  New _GotFocusAPI and _LostFocusAPI functions will be
required for each of PD's custom controls, because VB's internal focus
events are not reliable, but that's okay as it would be nice to have
these for other purposes.

4) Non-destructive changes (e.g. changing text layer settings) now
properly trigger enabling of the Save button.  Previously, you had to do
something destructive to make that button enable - serious bug!

5) The new system allows me to coalesce Undo entries under a single
header, with support for custom text modifications.  For example, under
the header "Modify text layer" I can now list the categorical change
made, e.g. "Modify text layer (style: bold)".  I hope to make this even
more fine-grained in the future, but some discussions with translators
are required as this does increase their burden.

I'm going to slowly roll this out to other text layer settings, and if
all goes well, I'll move generic layer opacity and blend mode here as
well.