Undo/Redo engine: implement event coalescing
If the same non-destructive event occurs twice in a row (e.g. changing
font once, then changing image zoom, then changing font again), the two
font changes can be coalesced into a single Undo event.

For non-destructive changes, this behavior is preferable as there's not
much merit to tracking multiple changes to the same non-destructive
setting.  This keeps the Undo/Redo stack tidy, and the Undo engine
handles it automatically (so long as the caller passes a flag notifying
it that coalescence is supported for the current action).