pdTextRenderer: fix left overhang on certain glyphs (particularly italics)
This was a pretty severe issue with e.g. Times New Roman, Italicized,
huge font size.  Both GDI and GDI+ were affected, because they always
assume an ABC width where A = 0.  Times New Roman, Italics, size 500
font has an A width of -24, or nearly 5% of the glyph width!

PD now manually calculates the left offset required to solve this issue,
as it varies depending on the initial glyph of the layer.  Both GDI+ and
GDI rendering paths have been fixed.