Text layers: integrate into macro recording/playback
Macro recording is such an ugly hack at present.  The whole engine
desperately needs an overhaul, but it's just not a major priority
compared to other tasks, do I'll have to settle for cobbling together
temporary fixes, while maintaining copious notes on what to improve in
the future.

With that in mind, this is mostly a "proof of concept" implementation
for the new non-destructive processor branch.  Text layer creation (and
non-destructive modifications!) can successfully be recorded and played
back on another image.  Usefulness will vary, as text positioning is
absolute, so playing back the macro on an image of a different size will
likely yield poor results.

Anyway, I know this is less likely to be helpful than non-destructive
layer changes (opacity and blend mode, in particular), but now that I
know the non-destructive processor pipeline works, I can start reworking
those changes to work with the pipeline.