pdComboBox_Font: start implementing modified management code
The font-specific variant of PD's combo box UC is going to be as
lightweight as possible.  There's no need for AddItem or RemoveItem
functions, as it simply queries the program's font cache for a list of
fonts, and displays them accordingly.

With this commit, list management, sizing, and display positioning has
all been implemented.  Still to-do is the actual "render a font sample
using each font" bit.