pdComboBox_Font: finished!
PD now provides a font-specific dropdown control, with support for font
previews.  I've implemented a custom caching system to avoid the need
for excessive font creation/destruction, and I think the final result is
pretty great!

I looked at a lot of different software when designing how fonts would
actually be "previewed".  I'm not a big fan of Microsoft's method of
rendering the font name itself in its own font.  Symbol fonts like
Wingdings make this method unpredictable and ugly, not to mention that
it's very difficult to compare fonts as each one is rendered with
different letters.

PhotoShop's method of columns is a little better, but I find it
aesthetically unappealing, and the tiny font sizes and sample text is
odd.

PD's method is to render the font name (in the program UI font) on the
left, and a live preview on the right.  Preview text is hard-coded, but
I really want to find a way to determine a font's preferred character
set, and use different preview text for different languages (e.g.
preview a Chinese font with Chinese characters, instead of English
ones), but I haven't figured out how to do it just yet.

In the meantime, PD's solution is fast, lightweight, and way less prone
to flicker than its competitors - so a win all around, I think.

Some clipping is possible if the preview font has enormous characters.
I'm debating whether to let the left margin of the preview text float,
if the font name is short enough to allow it.  This has some performance
implications as we have to manually measure each string, but it may look
a little prettier... I'll do some testing to see if it's viable.