pdTextRenderer: add extra padding to right margin...
...if left margin presents overhang.  The tough thing with right margins
is that we can't easily predict where GDI or GDI+ will choose to
wordwrap international text.  As such, all we can do is check for fonts
with known overhang, and estimate an amount of right-side padding that
will minimize clipping.

When we draw all our own glyphs, this problem won't exist, as we'll know
wordwrap positions in advance and can pad accordingly.