Start work on a Uniscribe interface
This is going to be a hideous project, but I've gotta start somewhere.

This commit is mostly an API copy+paste job to list basic Uniscribe
types, enums, and functions, modified to be VB-compatible, and allowing
PD to still compile.

As noted in the module header: "Many thanks to Michael Kaplan for his
endless work in demystifying Windows text handling.  Of particular value
to this module is this link from his personal blog:
http://www.siao2.com/2006/06/12/628714.aspx"