pdTextRenderer: switch to Uniscribe for all script shaping and layout
Add this to the "holy crap I can't believe this worked" pile.  ;)

To demonstrate the *awesome* text features PD can now provide, try the
following on Win 7 or later:

1) Create a new image with default settings
2) Switch to the typography tool (the fancy "t" icon)
3) Set the font to Segoe Script and choose a nice, big size (100 px)
4) Click-drag a large text layer, then type something like "we want all
the things"

If you watch the image as you type each character, you can see how the
scripting engine customizes the shape of each letter contextually,
depending on the surrounding characters.  Segoe Script is an OpenType
font with support for contextual swashes and alternates, and Uniscribe
uses that data to provide customization far beyond anything GDI or GDI+
can do.  (You can verify this by switching between the Text and
Typography tools, and seeing how the text changes when using the default
WAPI renderer.)

This is just the tip of the iceberg, and there are a ton of things still
to fix (like line breaks), but this is a significant landmark.

(Also, I'm pretty sure this is the only open-source implementation of
Uniscribe features in VB, so yay for PD!)