pdUniscribe: major refactoring to handle embedded runs
Man, this was a ton of work.

PD's custom glyph renderer can now handle intermixed LTR and RTL text
without trouble.  This is disastrously complex and it requires a ton of
unsafe pointer manipulation, but these are the hoops through which we
jump when interacting with complicated low-level APIs.

This work was also a prerequisite to proper line-breaking, which is a
prerequisite to implementing horizontal and vertical alignment options,
so it needed to be done.

With this commit, strings with missing glyphs (e.g. characters not
supported by the current font) are going to cause trouble.  I'm working
on an intermediate fallback until I have font linking working, but for
now, the output will be unpredictable gibberish.