pdUniscribe: retrieve wordbreak and whitespace identifiers
Hard linebreaks in the original string are still problematic, but with
this commit, we have everything we need for *automatic* word-wrapping
(which was a blocker for all remaining typography tool tasks, so yay for
finishing it!).

My main hold-up with hard line breaks is figuring out which class should
assume responsibility for them.  Uniscribe doesn't touch hard line
breaks.  (The paragraph is the largest unit of text it addresses.)
Theoretically, the caller is supposed to break strings into paragraphs
prior to interacting with Uniscribe.  After Uniscribe processes all
paragraphs, the caller then (somehow) manually merges the multiple
Uniscribe results into one conglomerate run suited for display.

This separation of responsibility isn't ideal for PD, as the text
renderer doesn't handle character-level details.  It just forwards the
relevant text to Uniscribe, then awaits a list of glyphs it can blindly
iterate.  Ideally, line-breaks would exist in that list of glyphs,
perhaps as specially marked "non-display" glyphs that are easily
recognized in the pre-rendering layout loop.

The only way to easily maintain this division of labor is to have PD's
Uniscribe class "inject" fake linebreak glyphs into the final glyph
collection, which is messy as there's not a 1:1 correlation between
characters (what the class receives) and glyphs (what the class
returns).

As I see it, I'll need to pre-scan the string, make a note of all hard
linebreak positions (and the usual Cr/Lf normalizing), replace
linebreaks with spaces, do all Uniscribe processing as if it's a
contiguous paragraph, then use the logical cluster information from
ScriptShape to figure out which glyphs correlate to which characters in
the original string, and use that to convert placeholder space glyphs
back into linebreak markers.

What a headache.  :/

That said, there are two big benefits to offloading linebreak handling
onto pdUniscribe:

1) When provided with a clipping width, pdUniscribe can return the glyph
collection already divided into lines.  This will make pdTextRenderer's
work much, much simpler.
2) It's a little unwieldy, but Uniscribe is capable of calculating
justified text positioning, which is awesome as most image editors are
restricted to left/center/right only.  (Justifying Arabic text in
particular is a nightmare, but Uniscribe can handle this via the
specialized ScriptJustify function.)  I wasn't planning on handling
fully justified alignment, but I may as well if we're going to be doing
other alignment stuff inside pdUniscribe.

Anyway, writing all this out is mostly for my own sake as I'm thinking
of tackling some rendering-specific bits before finishing PD's Uniscribe
interface, and I don't want to forget my plans in the interim!