Uniscribe: given a font name, retrieve supported scripts
This not only works, it works brilliantly and very quickly - much more
quickly than I anticipated!

Fonts are not required to explicitly list the scripts they support, so
this mechanism isn't foolproof (though it's really only a problem for
amateur fonts).  The only "foolproof" method, if such a thing exists, is
to manually probe Unicode character ranges, then map these to expected
script ranges.  That's incredibly cumbersome so I'm not apt to try it.