pdComboBox_Font: add preview support for CJK, Arabic, and Hebrew scripts
Many thanks to @Jonney3099 for his help with this.

PD's font selection box will now preferentially supply custom preview
text in the following order of script support:

CJK
Arabic
Hebrew
Latin
[Unknown scripts]

More scripts are forthcoming, but I wanted a good mix of ideographs and
RTL languages for this initial test.

The Latin chars "AaBbCc 123" are appended to all previews, as all fonts
are likely to support these base ASCII characters.

If a font supports any of the listed scripts, extra text specific to
that script is appended to the preview.  This is awesome for getting an
"at-a-glance" idea of what kind of Unicode coverage a font provides.

I've also revisited a bunch of the sizing decisions with the control.
The drop-down is now wider, with much more whitespace, and the font
previews are significantly larger than before.  This is important for
ideographic scripts, which need a lot of room to breathe.

Finally, to do this 100% correctly, I need to provide a UI where the
user can specify a language of choice.  This is relevant not just for
translations, but for knowing which script to preferentially preview in
the font dropdown.  If the user has a preferred script, we could always
display previews in that script (if supported by a font).

Until I've gotten that framework implemented, however, I think this is a
nice intermediary solution.