pdUniscribe: add support for hard linebreaks
I think this system will work out alright.  Unicode's ZWNJ character is
an easy replacement for hard linebreaks, and when it comes time to
return the glyph list to the user, we simply compare ZWNJ glyphs against
known linebreaks in the original string.  Matches are manually flagged
so the renderer doesn't have to do any extra detective work.

With this, I think everything is ready to start work on a new glyph
layout engine, which takes all this ridiculously complicated information
and pre-positions each glyph line.  Once that's ready, alignment will be
easy to calculate, and we can finally start moving on to fun features.