Tons of UI work for forthcoming typography features
Still tons more UI work to do (especially synchronizing everything
against the active layer object), but this is a good start.

Appearance options for this tool are plentiful, and a number of
additional user controls will be required (e.g. gradient, texture -
basically anything we'll reuse for brushes, shapes, and other vector
tools).

I'll get to those UCs in time, but for now, my goal is just to block out
a reasonable layout for this tool's many, many options, and start
assembling a to-do list for my GDI+ interface.