pdTextRenderer: add fallbacks if GDI+ fails during actual rendering
For reasons known only to the GDI+ gods, GDI+ may create a font just
fine, then arbitrarily fail when attempting to render with it.

This happens with the font "Batang", at size 197.1 or greater.  197
works just fine (which makes no sense), and a font of size 197.1 or
greater can be created without trouble, but attempting to render with it
will fail with "generic error."

GDI has no such problem, so as usual, it fills in if GDI+ decides to
randomly bail on us.