Typography tool: implement alignment
Horizontal and vertical alignments are now working, with a ton of extra
logic for correctly handling glyph overhang.  (As usual, most other
photo editors don't do this, so I'm happy to get it right!)

I'm not sure who bears responsibility for this, but some fonts have
slightly underreported ABC widths, which means minor edge clipping is
still a possibility.  I'll be adding manual margin settings to deal with
this.

(The end goal for this tool is to provide typography options roughly
comparable to CSS, so margins should be available anyway.)