pdComboBox UCs: placeholder commit against issue #171
Fixing the animation direction of the drop-down window has proven to be
a nightmare.  For reasons I have yet to solve, Windows resets the
drop-down size if the caller manually animates it, vs letting the system
animate it automatically.  I can't find a message to subclass that lets
me intercept and override this behavior, and I've wasted a
disproportionate amount of time trying to solve this.

So I'm saving my current work, which includes safely subclassing the
dropdown list box at dropdown time, so I can revisit in the future.