pdCanvas: overhaul tool interaction code
Also, get top-left layer transformations working.  Now that it's
complete, and all support code is in place, the other transforms are
going to be cake!

The old move/size tool code performed a ton of redundant calculations,
since it was created at a time when PD's canvas UC was just a thin
wrapper for image display.  Now that the canvas is much more capable, it
handles a lot of calculations so the caller doesn't have to - and the
move/size tool finally takes advantage of that.

A bunch of other tool rework was done to prep for paint tools, and the
canvas UC is slowly getting toward an acceptable state (although there's
a lot of refactoring still to do).