Myriad layer interaction improvements
- On-canvas layer transforms (by click-dragging layer corners) now work
for all corners, even while layer rotation is active.

- To simplify transformations, on-canvas transforms are now handled
relative to a layer's center.  So click-dragging any corner applies an
identical transformation, by sizing/moving the layer without changing
its current position.  I've seen this in other software and while it's a
little jarring relative to PD's past behavior, I think it's ultimately
more intuitive, especially given the way image interactions work on
mobile devices (which the average user is probably *more* familiar with
at this point).

- Obviously, you can still move a layer by clicking anywhere inside its
bounds and dragging it around.

- Shift key behavior is better.  I still need to fix its behavior if you
release the shift key before releasing the mouse, but while moving the
mouse, the use of Shift (to lock aspect ratio) behaves a lot more
naturally.

- New text layers created via click (and no subsequent drag) are now
sized to the union rect of the current viewport and the image itself.
This is much better than forcing new text layers to the size of the
image, especially if the user is currently zoomed-in.  (Note that you
can still click-drag to set your own text layer size at creation-time.)

- On-canvas nodes and layer boundaries have received some clean-up, so
line junctions and endings are now rendered as continuous shapes instead
of discrete lines.

- A new on-canvas node has been added for rotation.  It is rendered and
hit-detected correctly, but you can't otherwise interact with it just
yet.  (That's coming!)

Still WIP:

- Auto-detecting the layer under the mouse is still broken, but I've
finally reworked the tool code to a point where I can fix this (soon).

- The default viewport pipeline is still broken, so only the UI nodes
will rotate, not the image, unless you manually set the viewport
performance setting to "best quality".  I'll be fixing this soon as
well.