Add non-destructive x/y shear properties to pdLayer
I can't promise that these will actually work, but since I've gone to
the trouble of implementing matrix-based affine transforms, it should
theoretically be trivial to add non-destructive shearing too.

If my theory proves incorrect, this commit "causes no harm" so it won't
need to be rolled back (and perhaps I can revisit in the future).