Typography tool: finish last character-level features
(Well, maybe not the *absolute last* as I still need to fix underline
and strikethrough... sigh)

Anyway, after this commit, it's time to shift gears and start working on
dedicated UCs for pen and brush selection.  This will not only let me
clean up the Typography tool's various "visual" panels, but will also
save me a ton of work on eventual Shape tools (which make use of the
same fill and stroke features).