Add GetOutlineTextMetrics() support, so I can fix underline/strikeout behavior
GetOutlineTextMetrics is a neat WAPI function that provides a ton of
useful data, but man, struct alignment for the return is a disaster, and
MSDN isn't helpful because VB silently performs its own word alignment,
negating the need for some of the dummy padding bytes.

I think I've sorted everything out, since I'm getting proper returns on
the underline and strikeout entries on both XP and Win 8.  Fingers
crossed that more testing bears this out.