Initial build of fill user control + settings dialog
WARNING: massive commit!

PD 7.0 is going to provide a single master Fill class (brush in
GDI/GDI+) and Stroke class (pen in GDI/GDI+) as the back-end for
anything that can be filled and/or outlined, eg. text, shapes, arrows,
whatever.  My goal is to make these classes support every fill/outline
setting under the sun, and because the resulting settings are just
stored as strings, all new tools will automatically support every
fill/outline setting under the sun without any extra work on my part.

This commit is the first step toward making that possible.  The
Typography tool already exposed a lot of these features manually, but
now, I've reworked the fill-specific settings into a single, standalone
UC with a custom dialog and corresponding render class
(pdGraphicsBrush).  From here on out, any feature added to the user
control + rendering class will be automatically available to all other
tools, which is great!

There is still a *ton* to implement in the dialog, and textures in
particular are going to be weird and messy, but this is an important
first step.  Note that this commit *IS NOT BACKWARDS COMPATIBLE WITH
TEXT IN EXISTING PDI FILES*.  PDI text layers will still save/load just
fine, but any previous fill settings will be lost, as they are now
stored in a totally different format.  I will likely be modifying this
format further in the coming weeks, so as with any developer build,
don't save PDI files you can't afford to lose (just in case).  Note that
text background region fill settings are also lost; however, text
backgrounds now support 100% of the features available when filling the
text, so you now have way more options available for text backgrounds.

The outline/pen UC is still TBD, but now that the bulk of the
infrastructure for the Brush UC + rendering class is done (all that's
left is feature implementation), I may turn to the outline/pen tool so I
can call the Typography tool more or less done.  Then I can move onto
implementing other on-canvas tools while still plugging away at
fill/outline features.  We'll see.