Gradient editor: intelligently set color of new gradient nodes
This is made possible by a new "retrieve gradient color at arbitrary
position" function, which had to be implemented manually since we can't
use GDI+ to do it.  This function was also the key factor for using
gradients as bitmap effects a la gradient mapping, so we'll be using it
a lot in the coming days.