pdFSO: continued improvements
One thing to note: I've modified the class to always allow sharing
rights on created file handles (read, write, and delete/rename).  If the
IDE goes down, as it is prone to doing, unclosed file handles with
limited sharing rights prevent me from accessing key PD files without a
reboot.  It's easier to simply enable sharing rights on all file
handles, "just in case".

Also, a neat function I've made available is the ability to flag file
handles for temporary files.  Per MSDN, this changes the caching
behavior of the file, with Windows preferentially leaving the file in
the system cache with the assumption that it will be deleted before the
thread closes.  This can greatly improve performance on temp files
(which is handy for PD as we generate quite a few of them)!