Add new fallbacks for failed GDI+ blur operations
GDI+ is more prone to failure and bugs with each new Windows version, so
I need to start assuming that random failures are not just possible, but
potentially likely.

To that end, QuickBlur (which is used by many disparate functions) will
now test for multiple levels of GDI+ failure, and plug in our own native
blur operations as necessary.