Convert pdSelection and pdDIB to pdFSO (for Unicode awareness)
This was a large and complicated conversion, in part because pdSelection
and pdDIB engage in some messy interop when saving selection files.

Since I was here, I completely reworked the way these classes read and
write files to focus much more on performance.  A ton of intermediary
DIB and array work was dropped, and as much as possible, these classes
will now read and write relevant data directly into its final buffer(s).
The performance improvements are particularly noticeable when working
with raster selection data in the Undo/Redo chain.