First step toward fixing tab-key behavior on PD's various UCs
Relates to #172

The ultimate goal for tab order in PD is to generate it automatically at
run-time.  This would save me a lot of trouble with tab order breaking
when controls are rearranged/removed/inserted, and there's no reason we
can't easily sort controls top-to-bottom, right-to-left, and
automatically assign tab order accordingly.

But before I can do that, I want to clean up the way PD's UCs report tab
behavior.  Instead of always handling it internally, it is sometimes
beneficial to report an actual TabPress event.  (For example, if a UC
lives inside another UC, a TabPress should send focus to the next
control *within* that UC, not a separate UC entirely!)

To that end, I've updated the text box UC to have adjustable tab
behavior.  A property now controls whether tab presses raise a TabPress
event, or whether it forwards focus immediately.  Note that - by design
- this breaks tab forwarding when a text box embedded in a UC receives a
tab press.

(After some testing, I'll be moving the "forward focus" code completely
out of the UC and into a standalone module, so we don't have to repeat
that code in all UCs.)