pdFont: refactor to minimize resource usage and prep for UC overhaul
Rather than use separate labels and input controls throughout the
project, I will be reworking PD's various input controls to support a
"title" property.  This greatly reduces complexity (by cutting the
program's control count roughly in half), and it makes it much easier to
automate certain tasks.

Because pdFont is a crucial part of this, I wanted to clean it up a bit.
The use of a temporary DIB has been dropped in favor of a bare DC.  I've
also rewritten various measurement functions to work just fine, even if
the class isn't attached to a DC yet.