Overhaul plugin initialization
Apologies for the widespread list of files affected, but this is kind of
a strange commit.

Over the past year, it's gradually become clear that the Windows color
management engine is a hot mess.  A lot of crucial functionality is
spread across two versions of the engine - the old ICMS engine, which
dates back to the Windows 95-ME era, and the new WCS engine, introduced
by Vista - so you can't simply use one engine or the other.  Instead,
you have to intermix the two depending on what you're trying to
accomplish.

The WCS engine is also frustrating because it's designed around
Microsoft's custom color profile format, which absolutely nobody
(including Microsoft) uses.  ICC profiles are the standard, and you'll
never encounter an image that uses the stupid Microsoft format.  Worse
still, things like color space constants are straight-up faulty, but
they've been left as-is for backwards compatibility.  For example, it
recently came to my attention
(http://www.vbforums.com/showthread.php?742257-Color-Management-(ICC-Profile)-support-in-VB6-guide-and-sample-project&p=4923873&viewfull=1#post4923873)
that on certain versions of Windows, requesting a default sRGB profile
will actually return the custom default WCS profile set by the user.  To
get an sRGB profile, you have to use the opposite constant -
LCS_WINDOWS_COLOR_SPACE.  AAAARRRGGGGHHHH.

Anyway, an alternative solution is available.  LittleCMS
(https://github.com/mm2/Little-CMS) is a fantastic color management
engine, and it's used by everyone under the sun (e.g. Google, Mozilla,
GIMP, ImageMagick, MS Office on OSX), but integrating it into PD is
complicated, as color management is a very low-level concept that shows
up in a ton of different places.  But I've gotten it compiled on my
local PC, and am starting to run tests to figure out how best to make
use of it in PD.

Because of the complexity of the work, I'll be pursuing LittleCMS
integration in a separate branch, but prior to that, I wanted to knock
out some much-needed plugin management fixes.  Thus this commit.

Among other things, this commit makes it much easier to add new plugins
in the future, by centralizing a bunch of redundant plugin management
code.  Since I was here, I've also implemented a request from several
users to automatically fix faulty .zip file extractions.  Some .zip
programs (e.g. WinZip, which people still use for reasons I can't
comprehend) do not honor .zip folder structure by default, so PD's
plugins may get dumped into the core PD folder instead of staying in the
/App folder where they belong.  PD will now detect this, and rebuild the
folder structure accordingly.

Still to-do is integrating the new plugin management engine with the
Plugin Manager dialog (which currently implements redundant copies of
many plugin tasks).  But at least this will make it easier to work on a
separate LittleCMS branch in the meantime.